Create nginx virtual host if not exists

+++ scripts/serve-hhvm.sh
@@ -35,8 +35,12 @@ block="server {
 }
 "
 
-echo "$block" > "/etc/nginx/sites-available/$1"
-ln -fs "/etc/nginx/sites-available/$1" "/etc/nginx/sites-enabled/$1"
+if [ ! -f "/etc/nginx/sites-available/$1" ];
+    echo "$block" > "/etc/nginx/sites-available/$1"
+fi
+if [ ! -f "/etc/nginx/sites-enabled/$1" ];
+    ln -fs "/etc/nginx/sites-available/$1" "/etc/nginx/sites-enabled/$1"
+fi
 service nginx restart
 service php5-fpm restart
 service hhvm restart
